// src/app/models/FormattedSubmission.model.ts
/**
 * Created by LonelyEnvoy on 2017/8/31.
 */

export class FormattedSubmission {
  id: string;
  createTime: string;
  problemContestId: string;
  problemIndex: string;
  problemName: string;
  verdict: string;
  programmingLanguage: string;
  timeConsumed: string;
  memoryConsumed: string;
  passedTestCount: string;
  problemTags: string;
  submitter: string;
}